transform command: logger
