+++ core/src/main/java/com/taobao/arthas/core/command/logger/LoggerCommand.java
@@ -21,6 +21,7 @@ import com.alibaba.arthas.deps.org.slf4j.LoggerFactory;
 import com.taobao.arthas.common.IOUtils;
 import com.taobao.arthas.common.ReflectUtils;
 import com.taobao.arthas.core.command.Constants;
+import com.taobao.arthas.core.command.model.LoggerModel;
 import com.taobao.arthas.core.shell.command.AnnotatedCommand;
 import com.taobao.arthas.core.shell.command.CommandProcess;
 import com.taobao.arthas.core.util.ClassLoaderUtils;
@@ -154,9 +155,9 @@ public class LoggerCommand extends AnnotatedCommand {
         }
 
         if (result) {
-            process.write("Update logger level success.\n");
+            process.end(0, "Update logger level success.");
         } else {
-            process.write("Update logger level fail. Try to specify the classloader with the -c option. Use `sc -d CLASSNAME` to find out the classloader hashcode.\n");
+            process.end(-1, "Update logger level fail. Try to specify the classloader with the -c option. Use `sc -d CLASSNAME` to find out the classloader hashcode.");
         }
     }
 
@@ -194,97 +195,20 @@ public class LoggerCommand extends AnnotatedCommand {
 
             if (loggerTypes.contains(LoggerType.LOG4J)) {
                 Map<String, Map<String, Object>> loggerInfoMap = loggerInfo(classLoader, Log4jHelper.class);
-                String renderResult = renderLoggerInfo(loggerInfoMap, process.width());
-
-                process.write(renderResult);
+                process.appendResult(new LoggerModel(loggerInfoMap));
             }
 
             if (loggerTypes.contains(LoggerType.LOGBACK)) {
                 Map<String, Map<String, Object>> loggerInfoMap = loggerInfo(classLoader, LogbackHelper.class);
-                String renderResult = renderLoggerInfo(loggerInfoMap, process.width());
-
-                process.write(renderResult);
+                process.appendResult(new LoggerModel(loggerInfoMap));
             }
 
             if (loggerTypes.contains(LoggerType.LOG4J2)) {
                 Map<String, Map<String, Object>> loggerInfoMap = loggerInfo(classLoader, Log4j2Helper.class);
-                String renderResult = renderLoggerInfo(loggerInfoMap, process.width());
-
-                process.write(renderResult);
+                process.appendResult(new LoggerModel(loggerInfoMap));
             }
         }
 
-    }
-
-    private String renderLoggerInfo(Map<String, Map<String, Object>> loggerInfos, int width) {
-        StringBuilder sb = new StringBuilder(8192);
-
-        for (Entry<String, Map<String, Object>> entry : loggerInfos.entrySet()) {
-            Map<String, Object> info = entry.getValue();
-
-            TableElement table = new TableElement(2, 10).leftCellPadding(1).rightCellPadding(1);
-            TableElement appendersTable = new TableElement().rightCellPadding(1);
-
-            Class<?> clazz = (Class<?>) info.get(LoggerHelper.clazz);
-            table.row(label(LoggerHelper.name).style(Decoration.bold.bold()), label("" + info.get(LoggerHelper.name)))
-                            .row(label(LoggerHelper.clazz).style(Decoration.bold.bold()), label("" + clazz.getName()))
-                            .row(label(LoggerHelper.classLoader).style(Decoration.bold.bold()),
-                                            label("" + clazz.getClassLoader()))
-                            .row(label(LoggerHelper.classLoaderHash).style(Decoration.bold.bold()),
-                                            label("" + StringUtils.classLoaderHash(clazz)))
-                            .row(label(LoggerHelper.level).style(Decoration.bold.bold()),
-                                            label("" + info.get(LoggerHelper.level)));
-            if (info.get(LoggerHelper.effectiveLevel) != null) {
-                table.row(label(LoggerHelper.effectiveLevel).style(Decoration.bold.bold()),
-                                label("" + info.get(LoggerHelper.effectiveLevel)));
-            }
-
-            if (info.get(LoggerHelper.config) != null) {
-                table.row(label(LoggerHelper.config).style(Decoration.bold.bold()),
-                                label("" + info.get(LoggerHelper.config)));
-            }
-
-            table.row(label(LoggerHelper.additivity).style(Decoration.bold.bold()),
-                            label("" + info.get(LoggerHelper.additivity)))
-                            .row(label(LoggerHelper.codeSource).style(Decoration.bold.bold()),
-                                            label("" + info.get(LoggerHelper.codeSource)));
-
-            @SuppressWarnings("unchecked")
-            List<Map<String, Object>> appenders = (List<Map<String, Object>>) info.get(LoggerHelper.appenders);
-            if (appenders != null && !appenders.isEmpty()) {
-
-                for (Map<String, Object> appenderInfo : appenders) {
-                    Class<?> appenderClass = (Class<?>) appenderInfo.get(LoggerHelper.clazz);
-
-                    appendersTable.row(label(LoggerHelper.name).style(Decoration.bold.bold()),
-                                    label("" + appenderInfo.get(LoggerHelper.name)));
-                    appendersTable.row(label(LoggerHelper.clazz), label("" + appenderClass.getName()));
-                    appendersTable.row(label(LoggerHelper.classLoader), label("" + appenderClass.getClassLoader()));
-                    appendersTable.row(label(LoggerHelper.classLoaderHash),
-                                    label("" + StringUtils.classLoaderHash(appenderClass)));
-                    if (appenderInfo.get(LoggerHelper.file) != null) {
-                        appendersTable.row(label(LoggerHelper.file), label("" + appenderInfo.get(LoggerHelper.file)));
-                    }
-                    if (appenderInfo.get(LoggerHelper.target) != null) {
-                        appendersTable.row(label(LoggerHelper.target),
-                                        label("" + appenderInfo.get(LoggerHelper.target)));
-                    }
-                    if (appenderInfo.get(LoggerHelper.blocking) != null) {
-                        appendersTable.row(label(LoggerHelper.blocking),
-                                        label("" + appenderInfo.get(LoggerHelper.blocking)));
-                    }
-                    if (appenderInfo.get(LoggerHelper.appenderRef) != null) {
-                        appendersTable.row(label(LoggerHelper.appenderRef),
-                                        label("" + appenderInfo.get(LoggerHelper.appenderRef)));
-                    }
-                }
-
-                table.row(label("appenders").style(Decoration.bold.bold()), appendersTable);
-            }
-
-            sb.append(RenderUtil.render(table, width)).append('\n');
-        }
-        return sb.toString();
     }
 
     private static String helperClassNameWithClassLoader(ClassLoader classLoader, Class<?> helperClass) {
